feat(process): return history
