+++ examples/process_history/main.go
@@ -1,59 +1,61 @@
+// To make this example work make sure you have speech.ogg file in the root of directory
 package main
 
 import (
 	"context"
 	"fmt"
+	"os"
 
 	_ "github.com/joho/godotenv/autoload"
+	goopenai "github.com/sashabaranov/go-openai"
 
 	"github.com/neurocult/agency"
 	"github.com/neurocult/agency/providers/openai"
 )
 
 func main() {
-	provider := openai.New(openai.Params{Key: "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"})
-	params := openai.TextToTextParams{
-		Model:       "gpt-3.5-turbo",
-		Temperature: openai.Temperature(0),
-		MaxTokens:   100,
-	}
+	provider := openai.New(openai.Params{Key: os.Getenv("OPENAI_API_KEY")})
 
-	result, err := agency.NewProcess(
-		agency.ProcessStep{
-			Operation: provider.
-				TextToText(params).
-				SetPrompt("Increase the number by adding 1 to it. Answer only in numbers, without text"),
-		},
-		agency.ProcessStep{
-			Operation: provider.
-				TextToText(params).
-				SetPrompt("Double the number. Answer only in numbers, without text"),
-		},
-		agency.ProcessStep{
-			ConfigFunc: func(history agency.ProcessHistory, cfg *agency.OperationConfig) error {
-				firstStepResult, _ := history.Get(0)
-				cfg.Prompt = fmt.Sprintf("Add %s", firstStepResult)
-				return nil
-			},
-			Operation: provider.TextToText(params),
-		},
-	).Execute(
-		context.Background(),
-		agency.UserMessage("5"),
-		func(in, out agency.Message, cfg *agency.OperationConfig, stepIndex uint) {
-			fmt.Printf("---\n\nSTEP %d executed\n\nINPUT: %v\n\nCONFIG: %v\n\nOUTPUT: %v\n\n", stepIndex, in, cfg, out)
-		},
-	)
+	// step 1
+	hear := provider.
+		SpeechToText(openai.SpeechToTextParams{
+			Model: goopenai.Whisper1,
+		})
 
+	// step2
+	translate := provider.
+		TextToText(openai.TextToTextParams{
+			Model:       "gpt-3.5-turbo",
+			Temperature: openai.Temperature(0.5),
+		}).
+		SetPrompt("translate to russian")
+
+	// step 3
+	uppercase := provider.
+		TextToText(openai.TextToTextParams{
+			Model:       "gpt-3.5-turbo",
+			Temperature: openai.Temperature(1),
+		}).
+		SetPrompt("uppercase every letter of the text")
+
+	sound, err := os.ReadFile("speech.mp3")
 	if err != nil {
 		panic(err)
 	}
 
-	fmt.Println(result)
-}
+	ctx := context.Background()
+	speechMsg := agency.Message{Content: sound}
 
-// InjectHistory allows to pass history between operations by injecting it into the config.
-func InjectHistory(history agency.ProcessHistory, cfg *agency.OperationConfig) error {
-	cfg.Messages = history.All()
-	return nil
+	_, history, err := agency.ProcessFromOperations(
+		hear,
+		translate,
+		uppercase,
+	).Execute(ctx, speechMsg)
+	if err != nil {
+		panic(err)
+	}
+
+	for _, msg := range history.All() {
+		fmt.Println(msg.String())
+	}
 }
