+++ examples/speech_to_text_to_image/main.go
@@ -22,7 +22,7 @@ func main() {
 		panic(err)
 	}
 
-	msg, err := agency.ProcessFromOperations(
+	msg, _, err := agency.ProcessFromOperations(
 		factory.SpeechToText(openai.SpeechToTextParams{Model: goopenai.Whisper1}),
 		factory.TextToImage(openai.TextToImageParams{
 			Model:     goopenai.CreateImageModelDallE2,
